rust: change uint to usize for v1 compliance

+++ src/session.rs
@@ -25,10 +25,10 @@ impl Session {
 
     /// Counts the entities on a specific coordinate (in tile). Returns -1 if
     /// something is wrong with the given coordinates (out of bounds)
-    pub fn map_count_at(&self, coord: (uint, uint)) -> i32 {
+    pub fn map_count_at(&self, coord: (usize, usize)) -> i32 {
         let (x, y) = coord;
-        let x32 = x as i32;
-        let y32 = y as i32;
+        let x32 = x;
+        let y32 = y;
 
         // if self.map_height()    < y32
         //     || self.map_width() < x32
